[TUBDMP-54] More tweaks to demo mode
* ShibaLike now with only 1 user "dmp" with limited admin rights
* 2 different info boxes for dashboard & login-info
* Projects get approved automatically
* Removed tiny mess (the word "mappings") from admin dashboard
* No more user edit/delete
* Hidden data sources
* env variable is now boolean DEMO_MODE

+++ resources/views/admin/partials/user/info.blade.php
@@ -12,13 +12,15 @@
         @endif
     </td>
     <td>
-        {!! Form::open(['method' => 'GET', 'route' => ['admin.user.edit', $user->id], 'style'=>'display:inline-block', 'class' => 'form-inline', 'role' => 'form']) !!}
-        {!! Form::button('<i class="fa ' . trans('admin/table.icon.edit') . '"></i>', ['type' => 'submit', 'class' => 'btn btn-link btn-xs nopadding', 'title' => trans('admin/table.label.edit')] )  !!}
-        {!! Form::close() !!}
-        &nbsp;&nbsp;
-        {!! Form::open(['method' => 'DELETE', 'route' => ['admin.user.destroy', $user->id], 'style'=>'display:inline-block', 'class' => 'form-inline', 'role' => 'form']) !!}
-        {!! Form::button('<i class="fa ' . trans('admin/table.icon.delete') . '"></i>', ['type' => 'submit', 'class' => 'btn btn-link btn-xs nopadding', 'title' => trans('admin/table.label.delete')] )  !!}
-        {!! Form::close() !!}
+        @if (!env('DEMO_MODE'))
+            {!! Form::open(['method' => 'GET', 'route' => ['admin.user.edit', $user->id], 'style'=>'display:inline-block', 'class' => 'form-inline', 'role' => 'form']) !!}
+            {!! Form::button('<i class="fa ' . trans('admin/table.icon.edit') . '"></i>', ['type' => 'submit', 'class' => 'btn btn-link btn-xs nopadding', 'title' => trans('admin/table.label.edit')] )  !!}
+            {!! Form::close() !!}
+            &nbsp;&nbsp;
+            {!! Form::open(['method' => 'DELETE', 'route' => ['admin.user.destroy', $user->id], 'style'=>'display:inline-block', 'class' => 'form-inline', 'role' => 'form']) !!}
+            {!! Form::button('<i class="fa ' . trans('admin/table.icon.delete') . '"></i>', ['type' => 'submit', 'class' => 'btn btn-link btn-xs nopadding', 'title' => trans('admin/table.label.delete')] )  !!}
+            {!! Form::close() !!}
+        @endif
 
         {!! HTML::linkRoute('admin.user.projects.index', trans('admin/table.label.edit') . ' Projects', $user, ['class' => 'edit-relation']) !!}
     </td>
